Relocated UserService
