+++ api/src/models/customer/Customer.ts
@@ -14,7 +14,7 @@ import CustomerService from './CustomerService';
 import { PaginationWhereInput } from '../general/Pagination';
 import { UserConnection, UserCustomerType } from '../users/User';
 import DialogueService from '../questionnaire/DialogueService';
-import UserService from '../../services/user/UserService';
+import UserService from '../users/UserService';
 import isValidColor from '../../utils/isValidColor';
 
 export interface CustomerSettingsWithColour extends CustomerSettings {
